requirements.txt
timm==0.6.7
pillow==9.2.0
matplotlib==3.5.3
seaborn==0.12.2
torchmetrics==0.11.2
torchprofile==0.0.4
tqdm==4.65.0
wandb==0.15.5